requirements.txt
pandas==1.4.3
plotly==5.9.0
pyperclip==1.8.2
streamlit==1.11.0
Unidecode==1.3.4
scipy
datasets~=2.4.0
umap-learn
clarinpl-embeddings
pynndescent
flair~=0.11.3
numpy~=1.23.3
matplotlib~=3.5.3
seaborn~=0.12.0
scikit-learn~=1.1.2